server/alz-detect-api-hf/requirements.txt
fastapi[standard]
python-dotenv
uvicorn
pydantic
rich
python-multipart
tensorflow==2.15.0
keras==2.15.0
tf-keras==2.15.0
tensorflow-addons==0.23.0
numpy
pillow
vit-keras
typing-extensions
opencv-python-headless
matplotlib
